Move Contacts to separate files, create contact utils, add contact point rendering
