+++ newton/utils/__init__.py
@@ -13,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import numpy as np
 import warp as wp
 
 from .download_assets import clear_git_cache, download_asset
@@ -100,123 +99,6 @@ def vec_leaky_max(a: wp.vec3, b: wp.vec3):
 @wp.func
 def vec_abs(a: wp.vec3):
     return wp.vec3(wp.abs(a[0]), wp.abs(a[1]), wp.abs(a[2]))
-
-
-def load_mesh(filename: str, method: str | None = None):
-    """
-    Loads a 3D triangular surface mesh from a file.
-
-    Args:
-        filename (str): The path to the 3D model file (obj, and other formats supported by the different methods) to load.
-        method (str): The method to use for loading the mesh (default None). Can be either `"trimesh"`, `"meshio"`, `"pcu"`, or `"openmesh"`. If None, every method is tried and the first successful mesh import where the number of vertices is greater than 0 is returned.
-
-    Returns:
-        Tuple of (mesh_points, mesh_indices), where mesh_points is a Nx3 numpy array of vertex positions (float32),
-        and mesh_indices is a Mx3 numpy array of vertex indices (int32) for the triangular faces.
-    """
-    import os
-
-    if not os.path.exists(filename):
-        raise FileNotFoundError(f"File not found: {filename}")
-
-    def load_mesh_with_method(method):
-        if method == "meshio":
-            import meshio
-
-            m = meshio.read(filename)
-            mesh_points = np.array(m.points)
-            mesh_indices = np.array(m.cells[0].data, dtype=np.int32)
-        elif method == "openmesh":
-            import openmesh
-
-            m = openmesh.read_trimesh(filename)
-            mesh_points = np.array(m.points())
-            mesh_indices = np.array(m.face_vertex_indices(), dtype=np.int32)
-        elif method == "pcu":
-            import point_cloud_utils as pcu
-
-            mesh_points, mesh_indices = pcu.load_mesh_vf(filename)
-            mesh_indices = mesh_indices.flatten()
-        else:
-            import trimesh
-
-            m = trimesh.load(filename)
-            if hasattr(m, "geometry"):
-                # multiple meshes are contained in a scene; combine to one mesh
-                mesh_points = []
-                mesh_indices = []
-                index_offset = 0
-                for geom in m.geometry.values():
-                    vertices = np.array(geom.vertices, dtype=np.float32)
-                    faces = np.array(geom.faces.flatten(), dtype=np.int32)
-                    mesh_points.append(vertices)
-                    mesh_indices.append(faces + index_offset)
-                    index_offset += len(vertices)
-                mesh_points = np.concatenate(mesh_points, axis=0)
-                mesh_indices = np.concatenate(mesh_indices)
-            else:
-                # a single mesh
-                mesh_points = np.array(m.vertices, dtype=np.float32)
-                mesh_indices = np.array(m.faces.flatten(), dtype=np.int32)
-        return mesh_points, mesh_indices
-
-    if method is None:
-        methods = ["trimesh", "meshio", "pcu", "openmesh"]
-        for method in methods:
-            try:
-                mesh = load_mesh_with_method(method)
-                if mesh is not None and len(mesh[0]) > 0:
-                    return mesh
-            except Exception:
-                pass
-        raise ValueError(f"Failed to load mesh using any of the methods: {methods}")
-    else:
-        mesh = load_mesh_with_method(method)
-        if mesh is None or len(mesh[0]) == 0:
-            raise ValueError(f"Failed to load mesh using method {method}")
-        return mesh
-
-
-def visualize_meshes(
-    meshes: list[tuple[list, list]], num_cols=0, num_rows=0, titles=None, scale_axes=True, show_plot=True
-):
-    # render meshes in a grid with matplotlib
-    import matplotlib.pyplot as plt
-
-    if titles is None:
-        titles = []
-
-    num_cols = min(num_cols, len(meshes))
-    num_rows = min(num_rows, len(meshes))
-    if num_cols and not num_rows:
-        num_rows = int(np.ceil(len(meshes) / num_cols))
-    elif num_rows and not num_cols:
-        num_cols = int(np.ceil(len(meshes) / num_rows))
-    else:
-        num_cols = len(meshes)
-        num_rows = 1
-
-    vertices = [np.array(v).reshape((-1, 3)) for v, _ in meshes]
-    faces = [np.array(f, dtype=np.int32).reshape((-1, 3)) for _, f in meshes]
-    if scale_axes:
-        ranges = np.array([v.max(axis=0) - v.min(axis=0) for v in vertices])
-        max_range = ranges.max()
-        mid_points = np.array([v.max(axis=0) + v.min(axis=0) for v in vertices]) * 0.5
-
-    fig = plt.figure(figsize=(12, 6))
-    for i, (vertices, faces) in enumerate(meshes):
-        ax = fig.add_subplot(num_rows, num_cols, i + 1, projection="3d")
-        if i < len(titles):
-            ax.set_title(titles[i])
-        ax.plot_trisurf(vertices[:, 0], vertices[:, 1], vertices[:, 2], triangles=faces, edgecolor="k")
-        if scale_axes:
-            mid = mid_points[i]
-            ax.set_xlim(mid[0] - max_range, mid[0] + max_range)
-            ax.set_ylim(mid[1] - max_range, mid[1] + max_range)
-            ax.set_zlim(mid[2] - max_range, mid[2] + max_range)
-    if show_plot:
-        plt.show()
-    return fig
 
 
 __all__ = [
